+++ CHANGELOG.md
@@ -29,6 +29,7 @@ jsx-slack v2 has improved JSX structure and built-in components to output the re
 - HTML-compatible `<Option selected>` and `<RadioButton checked>`
 - `value` prop as an alias into `initialXXX` prop in some interactive components
 - Added JSDoc to many public APIs and components
+- Support new JSX transpile via `automatic` runtime in Babel >= 7.9 ([#142](https://github.com/speee/jsx-slack/pull/142))
 
 ### Removed